Revert "Merge branch 'main' of https://github.com/kartverket/geogpt into GEON-1679-Katalog-vis-flere-dataset"
This reverts commit d7c5165c21e203388216f62ca9848fdc55acabe3, reversing
changes made to b7dfdf0a0f7101614acc7b5afb755e30862d2774.

consistent ui fullskjer
