+++ geonorge-app/components/ui/message-list.tsx
@@ -60,8 +60,8 @@ export function MessageList({
                   variant="secondary"
                   className={`transition-colors relative ${
                     isValidWmsUrl(message.wmsUrl)
-                      ? "hover:bg-green-600 bg-green-500 text-white hover:text-white"
-                      : "bg-gray-300 text-gray-500 cursor-not-allowed"
+                      ? " rounded-[2px] bg-[#FF8B65] hover:bg-[#FE642F] text-white"
+                      : "rounded-[2px] bg-gray-300 text-gray-500 cursor-not-allowed"
                   }`}
                   disabled={!isValidWmsUrl(message.wmsUrl)}
                 >
@@ -71,7 +71,7 @@ export function MessageList({
                   <Button
                     onClick={() => onDownloadClick?.(message.downloadUrl!)}
                     variant="secondary"
-                    className="hover:bg-green-600 bg-green-500 text-white hover:text-white transition-colors relative"
+                    className="rounded-[2px] bg-[#404041] hover:bg-[#5c5c5d] text-white"
                   >
                     Last ned datasett
                   </Button>
