--- geonorge-server/src/llm/__init__.py
@@ -1,7 +1,0 @@
-"""
-LLM module for managing language model instances and related functionality.
-"""
-
-from .llmManager import LLMManager
-
-__all__ = ["LLMManager"]
