--- geonorge-server/src/rag/__init__.py
@@ -1,12 +1,0 @@
-"""
-Retrieval Augmented Generation (RAG) module for GeoNorge.
-"""
-
-from .chain import EnhancedGeoNorgeRAGChain
-from .response_handlers import get_rag_response, get_rag_context
-
-__all__ = [
-    'EnhancedGeoNorgeRAGChain',
-    'get_rag_response',
-    'get_rag_context'
-]
