--- geonorge-server/src/rag/models/__init__.py
@@ -1,7 +1,0 @@
-"""
-Models module for managing conversation state and related functionality.
-"""
-
-from .state import ConversationState
-
-__all__ = ["ConversationState"]
