--- geonorge-server/src/retrieval/__init__.py
@@ -1,7 +1,0 @@
-"""
-Retrieval module for handling document retrieval and processing from various data sources.
-"""
-
-from .geo_vector_retriever import GeoNorgeVectorRetriever
-
-__all__ = ["GeoNorgeVectorRetriever"]
